telegram_bot/requirements.txt
aiogram
openai
python-dotenv
httpx


